Use a common header file for Windows errno names.
In earlier Visual Studio and MinGW editions, BSD socket errno:s were
left undefined. This is no longer the case. This may cause build or
runtime errors. To avoid this, we are using a common header file which
corrects all definitions. This header will also be used with other
projects such as sercd, TigerVNC, PulseAudio etc.



git-svn-id: svn://svn.code.sf.net/p/unfs3/code/trunk@495 2d7181c2-f438-4696-8f62-672025efbd90